Files for client side of client server comms
These files are part of the overall client side application. As such the games will use these in order to communicate with the server or maintain local data on what games are happening on the network. See the files uploaded under Go in order to see how these communications are utilized.