These are files necessary for the operation of go
goboard holds a method that can display the go board. goboard also has a 2d array of all the go class objects - characterizes boardstate
go is a class that contains all the game mechanics for the go game. Each go class object points to one tile on the board. There are methods to send the board state description via the Client class to other players over the server.